refactor betrayal to multiple components
